src/third_party/blink/web_tests/external/wpt/wake-lock/wakelock-type.https-expected.txt

This is a testharness.js-based test.
FAIL Test that wakeLock.type is 'screen' when screen wake lock is invoked Illegal constructor
FAIL Test that wakeLock.type is 'system' when system wake lock is invoked Illegal constructor
PASS 'TypeError' is thrown when set an empty wake lock type
PASS 'TypeError' is thrown when set an invalid wake lock type
Harness: the test ran to completion.
